project/requirements.txt
asgiref==3.7.2
Django==4.2.11
djangorestframework==3.14.0
drf-base64==2.0
drf-yasg==1.21.7
gunicorn==20.1.0
inflection==0.5.1
packaging==23.2
pillow==10.2.0
pytz==2024.1
PyYAML==6.0.1
sorl-thumbnail==12.10.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
psycopg2-binary==2.8.6
